temp4
Found out that GameOverlay.Net couldn't put overlay on full-screen
program. Seems I should do good'ol P/Invoke thingy, but there's a
library which do it automatically for me.

btw this is not completed work and most of the commits are
to transfer my work between devices. It will be cleaned up
after some trial and error.